Update size_t to int

+++ sort.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-void insertion_sort(int array[], size_t length)
+void insertion_sort(int array[], int length)
 {
     for (int j = 1; j < length; ++j)
     {
