+++ sort.h
@@ -1,3 +1,3 @@
 #include <stdio.h>
 
-void insertion_sort(int [], size_t);
+void insertion_sort(int [], int);
