+++ util.c
@@ -5,17 +5,17 @@ void print_sort_header(const char* header)
     printf("===================== %s =====================\n", header);
 }
 
-void print_array(int array[], size_t length)
+void print_array(int array[], int length)
 {
     printf("[");
 
-    for (size_t i = 0; i < length; ++i)
+    for (int i = 0; i < length; ++i)
         printf("%d, ", array[i]);
 
     printf("\b\b]\n");
 }
 
-void print_sort(int array[], size_t length, void (*sort)(int [], size_t))
+void print_sort(int array[], int length, void (*sort)(int [], int))
 {
     printf("\tArray before sorting:\t");
     print_array(array, length);
